Use pkg-config to get includes and ld flags
The location of headers and libraries varies on different systems.
pkg-config gets the right path for the user's system.

Also, pkg-config --libs correctly gets dependencies like -lm and -lrt so
that we don't have to maintain that list manually.

Finally, I moved the `all` make target to be the first rule, which is
the usual convention. The first rule in a Makefile is the default, so
now if you type just `make`, it will run the `all` target.